# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Water Tracker APP",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "API for tracking the amount of water"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://h2oflow-team4-backend.onrender.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/auth/reset-password": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Password reset page",
        "operationId": "resetPasswordPage",
        "description": "Displays an HTML page for password reset",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Password reset token"
          }
        ],
        "responses": {
          "200": {
            "description": "HTML page for password reset",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "example": "<html><body>...</body></html>"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset user password",
        "operationId": "resetPassword",
        "description": "Resets the user password using a valid reset token",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "New password and reset token",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1..."
                  },
                  "password": {
                    "type": "string",
                    "example": "NewSecurePassword123!"
                  }
                },
                "required": [
                  "token",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Password reset successfully"
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register User",
        "operationId": "registerUser",
        "description": "Register user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/request-body-for-auth"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/register-201"
          },
          "409": {
            "$ref": "#/components/responses/register-409"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Login user",
        "operationId": "loginUser",
        "description": "Logs in a user and returns an access token. Refresh token is set in cookies.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/request-body-for-auth"
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully logged in an user!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "User logout",
        "operationId": "logoutUser",
        "description": "Logs out the user and removes session data from cookies and database",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully logged out"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user profile",
        "operationId": "getUser",
        "description": "Returns the profile data of the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/getUser-200"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update user profile",
        "operationId": "updateUser",
        "description": "Allows authenticated users to update their profile information",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "User profile update data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "description": "User name",
                    "type": "string",
                    "example": "Alex Jonson"
                  },
                  "email": {
                    "description": "User email",
                    "type": "string",
                    "example": "jonson_a@gmail.com"
                  },
                  "gender": {
                    "description": "User gender",
                    "type": "string",
                    "enum": [
                      "woman",
                      "man"
                    ],
                    "example": "man"
                  },
                  "weight": {
                    "description": "User weight in kilograms",
                    "type": "number",
                    "example": 67
                  },
                  "dailySportTime": {
                    "description": "User sport time in hours per day",
                    "type": "number",
                    "example": 8
                  },
                  "dailyNorm": {
                    "description": "Daily water intake in milliliters",
                    "type": "number",
                    "example": 1700
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/updateUser-200"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "400": {
        "description": "Bad request: Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Validation failed"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "User not found!",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "User not found!"
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Something went wrong"
                },
                "data": {
                  "type": "string",
                  "example": "Error details or message from the server"
                }
              }
            }
          }
        }
      },
...
      "register-201": {
        "description": "User successfully registered",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 201
                },
                "message": {
                  "type": "string",
                  "example": "Successfully registered a user!"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/user"
                    },
                    "accessToken": {
                      "type": "string",
                      "example": "abc123xgf4b62bYVrd53s565nt225"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "register-409": {
        "description": "Email already in use",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 409
                },
                "message": {
                  "type": "string",
                  "example": "Email in use"
                }
              }
            }
          }
        }
      },
...
      "getUser-200": {
        "description": "User successfully retrieved",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User successfully found!"
                },
                "user": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          }
        }
      },
      "updateUser-200": {
        "description": "User successfully updated",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User successfully updated!"
                },
                "user": {
                  "$ref": "#/components/schemas/user"
                }
              }
            }
          }
        }
      },
...
    },
    "requestBodies": {
      "request-body-for-auth": {
        "description": "User credentials",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "examplePsw123!"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "user": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "_id": {
            "description": "User id",
            "type": "string",
            "example": "83cf00c8706c3609f8a91ae9"
          },
          "name": {
            "description": "User name",
            "type": "string",
            "example": "Alex Jonson",
            "default": "User"
          },
          "email": {
            "description": "User email",
            "type": "string",
            "example": "jonson_a@gmail.com"
          },
          "gender": {
            "description": "User gender",
            "type": "string",
            "enum": [
              "woman",
              "man"
            ],
            "example": "man",
            "default": "woman"
          },
          "weight": {
            "description": "User weight in kilograms",
            "type": "number",
            "example": 67,
            "default": 0
          },
          "dailySportTime": {
            "description": "User sport time in hours per day",
            "type": "number",
            "example": 8,
            "default": 0
          },
          "dailyNorm": {
            "description": "Daily water intake in milliliters",
            "type": "number",
            "example": 1700,
            "default": 1500
          },
          "avatarUrl": {
            "description": "URL to the user profile photo",
            "type": "string",
            "example": "https://res.cloudinary.com/ddtrmtbl8/image/upload/v1740747328/qbclx8i3i9zchrvi1git.jpg",
            "default": "https://collection.cloudinary.com/dojuslnjs/d637fb90dcf69166ae1e1c04c01fb63e?"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-10T14:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-03-10T15:00:00.000Z"
          }
        }
      },
...
    }
  }
}